requirements.txt
streamlit>=1.35.0
sqlalchemy>=2.0.0
pandas>=2.0.0
fpdf2>=2.8.0
pytest>=7.0.0
